refactor(ts): move more types into server_api
